+++ R/combineFeatures.R
@@ -200,7 +200,7 @@ combineMatrixFeatures <- function(matr,    ## matrix
                                          return(medpol$overall + medpol$col)
                                      })
         } else if (method == "robust") {
-            summarisedFeatures <- by(matr, groupBy, robustSummary, ...)
+            summarisedFeatures <- by(matr, groupBy, MsCoreUtils::robustSummary, ...)
         } else if (method == "weighted.mean") {
             ## Expecting 'w' argument
             args <- list(...)
@@ -290,65 +290,4 @@ aggvar <- function(object, groupBy, fun) {
     ans <- cbind(agg_dist = d, nb_feats = nr)
     attr(ans, "agg_dist") <- fun
     ans
-}
-
-##' This function calculates the robust summarisation for each feature
-##' (protein). Note that the function assumes that the intensities in
-##' input `e` are already log-transformed.
-##'
-##' @title Calculate robust expression summary
-##' @param e A feature (peptide or spectra) by sample `matrix`
-##'     containing the expression data.
-##' @param ... Additional parameters passed to `MASS::rlm`
-##' @return `numeric()` vector of length `length(expression)` with
-##'     robust summarised values.
-##' @author Adriaan Sticker, Sebastian Gibb and Laurent Gatto
-##' @md
-##' @noRd
-robustSummary <- function(e, residuals = FALSE, ...) {
-    ## If there is only one 1 peptide for all samples return
-    ## expression of that peptide
-    if (nrow(e) == 1L) return(e)
-
-    ## remove missing values
-    p <- !is.na(e)
-    expression <- e[p] ## expression becomes a vector
-    sample <- rep(colnames(e), each = nrow(e))[p]
-    feature <- rep(rownames(e), times = ncol(e))[p]
-
-    ## model.matrix breaks on factors with 1 level so make vector of
-    ## ones (intercept).
-    if (length(unique(sample)) == 1L) sample <- rep(1, length(sample))
-
-    ## Sum contrast on peptide level so sample effect will be mean
-    ## over all peptides instead of reference level.
-    X <- stats::model.matrix(~ -1 + sample + feature,
-                             contrasts.arg = list(feature = 'contr.sum'))
-    ## MASS::rlm breaks on singulare values.
-    ## - Check with base lm if singular values are present.
-    ## - If so, these coefficients will be zero, remove this collumn
-    ##   from model matrix
-    ## - Rinse and repeat on reduced modelmatrx till no singular
-    ##   values are present
-    repeat {
-        fit <- stats::.lm.fit(X, expression)
-        id <- fit$coefficients != 0
-        X <- X[ , id, drop = FALSE]
-        if (!any(!id)) break
-    }
-    ## Last step is always rlm: calculate estimated effects effects as
-    ## summarised values
-    fit <- MASS::rlm(X, expression, ...)
-
-    sampleid <- seq_along(unique(sample))
-    ## This will be needed for NUSE-type of quality control, but will
-    ## need to check for missing data as below.
-    ## se <- unique(summary(fit)$coefficients[sampleid, 'Std. Error'])
-
-    ## Take the sample coefficients ( = summarised expression values)
-    coef  <-  fit$coefficients[sampleid]
-    ## Sort the sample coefficients in the same way as the samplenames
-    ## of expression matrix. Puts NA for the samples without any
-    ## expression value
-    coef[paste0('sample', colnames(e))]
 }